requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.12.1
flake8==3.9.2
idna==3.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
